+++ packages/infinity-modules/model-management/src/model-management-store.js
@@ -300,7 +300,7 @@ export default ({
     getRecordsByTagData: async ({ commit, dispatch }, payload) => {
       const data = await dispatch(
         'element/getRecordsByTags',
-        { payload },
+        payload,
         { root: true },
       );
       if (data) {