+++ packages/infinity-modules/report-viewer/src/components/ReportGrid.vue
@@ -1,4 +1,10 @@
 <template>
+<div>
+  <v-card
+    flat
+    class="transparent"
+    id="chart"
+  ></v-card>
   <ag-grid-vue
     :sideBar="true"
     :rowData="rowData"
@@ -10,7 +16,8 @@
     rowGroupPanelShow="always"
     :gridOptions="gridOptions"
     :enableRangeSelection="true"
-    class="ag-theme-balham mt-2"
+    :class="`${agGridTheme} mt-2`"
+    :localeText="agGridLocaleText"
     :defaultColDef="defaultColDef"
     style="width: 100%; height: 450px;"
     @sort-changed="onStateChange"
@@ -19,15 +26,19 @@
     @column-visible="onStateChange"
     @column-resized="onStateChange"
     @column-moved="onStateChange"
-    @column-row-group-changed="onStateChange"
+    @column-pivot-mode-changed="onPivotModeChange"
     @column-pivot-changed="onStateChange"
+    @column-row-group-changed="onStateChange"
     @column-value-changed="onStateChange"
+    @first-data-rendered="visualizeData"
   ></ag-grid-vue>
+</div>
 </template>
 
 <script>
 import 'ag-grid-community/dist/styles/ag-grid.css';
 import 'ag-grid-community/dist/styles/ag-theme-balham.css';
+import 'ag-grid-community/dist/styles/ag-theme-balham-dark.css';
 import { AgGridVue } from 'ag-grid-vue';
 import { mapState, mapMutations, mapGetters } from 'vuex';
 
@@ -39,6 +50,7 @@ export default {
   data() {
     return {
       rowData: [],
+      aggFunc: null,
       gridApi: null,
       columnDefs: [],
       gridOptions: null,
@@ -48,7 +60,9 @@ export default {
     };
   },
   created() {
-    this.gridOptions = {};
+    this.gridOptions = {
+      groupDefaultExpanded: -1,
+    };
     this.defaultColDef = {
       filter: true,
       sortable: true,
@@ -64,14 +78,19 @@ export default {
     this.gridColumnApi = this.gridOptions.columnApi;
   },
   computed: {
+    ...mapState('helper', ['isDark']),
+    ...mapGetters('helper', ['agGridLocaleText', 'agGridTheme']),
     ...mapState('reports', ['report', 'reportMapping']),
     ...mapGetters('reports', ['isBaseReport', 'gridObject', 'exportFileName']),
   },
   watch: {
     report(val) {
+      if (val) {
+        this.aggFunc = val.aggFunc || null;
+      }
       if (val && val.cols) {
         this.columnDefs = val.cols.map((col) => ({
-          headerName: col.description,
+          headerName: this.getHeaderName(col),
           field: col.name,
           colId: col.name,
           filter: this.getColumnFilter(col),
@@ -79,11 +98,73 @@ export default {
       }
       if (val && val.reportData) {
         this.rowData = val.reportData;
+        this.visualizeData();
       }
     },
   },
   methods: {
     ...mapMutations('reports', ['setGridState']),
+    getHeaderName(col) {
+      switch (this.$i18n.locale) {
+        case 'zhHans':
+          return col.description_cn || col.description;
+        case 'hi':
+          return col.description_hi || col.description;
+        case 'th':
+          return col.description_th || col.description;
+        case 'de':
+          return col.description_de || col.description;
+        default:
+          return col.description;
+      }
+    },
+    visualizeData() {
+      const chartContainer = document.getElementById('chart');
+      chartContainer.innerHTML = '';
+      chartContainer.style.height = 0;
+      if (this.gridColumnApi.isPivotMode()) {
+        this.createPivotChart(chartContainer);
+      } else {
+        this.createRangeChart(chartContainer);
+      }
+      chartContainer.style.height = '350px';
+    },
+    createRangeChart(chartContainer) {
+      const param = {
+        chartType: 'column',
+        cellRange: {
+          columns: this.columnDefs.map((c) => c.field),
+        },
+        aggFunc: this.aggFunc,
+        chartThemeOverrides: {
+          common: {
+            title: {
+              enabled: false,
+            },
+            legend: { enabled: true },
+            navigator: { enabled: true },
+          },
+        },
+        chartContainer,
+      };
+      this.gridApi.createRangeChart(param);
+    },
+    createPivotChart(chartContainer) {
+      const param = {
+        chartType: 'column',
+        chartThemeOverrides: {
+          common: {
+            title: {
+              enabled: false,
+            },
+            legend: { enabled: true },
+            navigator: { enabled: true },
+          },
+        },
+        chartContainer,
+      };
+      this.gridApi.createPivotChart(param);
+    },
     getColumnFilter(col) {
       const type = col && col.type.toLowerCase();
       switch (type) {
@@ -102,15 +183,18 @@ export default {
     onStateChange() {
       const colState = this.gridColumnApi.getColumnState();
       const groupState = this.gridColumnApi.getColumnGroupState();
-      const sortState = this.gridApi.getSortModel();
       const filterState = this.gridApi.getFilterModel();
+      const isPivotMode = this.gridColumnApi.isPivotMode();
       const state = {
         colState,
         groupState,
-        sortState,
         filterState,
+        isPivotMode,
       };
       this.setGridState(JSON.stringify(state));
+    },
+    onPivotModeChange() {
+      this.visualizeData();
     },
     restoreState() {
       if (!this.isBaseReport) {
@@ -123,14 +207,14 @@ export default {
     setState(state) {
       this.gridColumnApi.setColumnState(state.colState);
       this.gridColumnApi.setColumnGroupState(state.groupState);
-      this.gridApi.setSortModel(state.sortState);
       this.gridApi.setFilterModel(state.filterState);
+      this.gridColumnApi.setPivotMode(state.isPivotMode);
     },
     resetState() {
       this.gridColumnApi.resetColumnState();
       this.gridColumnApi.resetColumnGroupState();
-      this.gridApi.setSortModel(null);
       this.gridApi.setFilterModel(null);
+      this.gridColumnApi.setPivotMode(false);
     },
     exportGridCSV() {
       const params = {