+++ packages/infinity/src/store/modules/element.js
@@ -26,9 +26,13 @@ export default ({
       return false;
     },
 
-    getRecordsByTags: async (_, { payload }) => {
+    getRecordsByTags: async (_, { elementName, queryParam, request }) => {
       try {
-        const { data } = await ElementService.getRecordsByTags(payload);
+        const { data } = await ElementService.getRecordsByTags(
+          elementName,
+          queryParam,
+          request,
+        );
         if (data && data.results) {
           return data;
         }
@@ -38,6 +42,7 @@ export default ({
       return false;
     },
 
+<<<<<<< HEAD
     postStreamRecords: async (_, { payload }) => {
       try {
         const { data } = await ElementService.postStreamRecords(payload);
@@ -51,9 +56,24 @@ export default ({
     },
 
     createElement: async ({ dispatch }, element) => {
+=======
+    createElement: async ({ dispatch }, {
+      element,
+      webhooks = [],
+    }) => {
+>>>>>>> dev
       try {
         const { data } = await ElementService.createElement(element);
         if (data && data.results) {
+          for (let i = 0; i < webhooks.length; i += 1) {
+            // eslint-disable-next-line
+            await dispatch('createWebhook', {
+              payload: {
+                ...webhooks[i],
+                elementId: data.elementId,
+              },
+            });
+          }
           return data.elementId;
         }
         const elem = await dispatch('getElement', element.elementName);
@@ -78,10 +98,11 @@ export default ({
     createElementAndTags: async ({ dispatch }, {
       element,
       tags,
+      webhooks = [],
     }) => {
       let tagsCreated = false;
       try {
-        const elementId = await dispatch('createElement', element);
+        const elementId = await dispatch('createElement', { element, webhooks });
         if (elementId) {
           const tagsToProvision = tags.map((tag) => ({
             ...tag,
@@ -100,11 +121,21 @@ export default ({
       tags,
       records,
       assetId,
+      webhooks = [],
     }) => {
       let recordsCreated = false;
       try {
-        const elementId = await dispatch('createElement', element);
+        const elementId = await dispatch('createElement', { element, webhooks });
         if (elementId) {
+          for (let i = 0; i < webhooks.length; i += 1) {
+            // eslint-disable-next-line
+            await dispatch('createWebhook', {
+              payload: {
+                ...webhooks[i],
+                elementId,
+              },
+            });
+          }
           const tagsToProvision = tags.map((tag) => ({
             ...tag,
             elementId,
@@ -135,7 +166,7 @@ export default ({
     }) => {
       let upsert = false;
       try {
-        const elementId = await dispatch('createElement', element);
+        const elementId = await dispatch('createElement', { element });
         if (elementId) {
           const tagsToProvision = tags.map((tag) => ({
             ...tag,
@@ -173,10 +204,11 @@ export default ({
       tags,
       records,
       assetId = 0,
+      webhooks = [],
     }) => {
       let upsert = false;
       try {
-        const elementId = await dispatch('createElement', element);
+        const elementId = await dispatch('createElement', { element, webhooks });
         if (elementId) {
           const tagsToProvision = tags.map((tag) => ({
             ...tag,