+++ packages/infinity/vue.config.js
@@ -5,14 +5,14 @@ module.exports = {
   devServer: {
     proxy: {
       '/server': {
-        target: 'http://192.168.1.126/server',
+        target: 'http://192.168.1.161/server',
         changeOrigin: true,
         pathRewrite: {
           '^/server': '',
         },
       },
       '/sse': {
-        target: 'http://192.168.1.126/sse',
+        target: 'http://192.168.1.161/sse',
         changeOrigin: true,
         pathRewrite: {
           '^/sse': '',