+++ packages/origin-modules/customer-add-new/src/data/webappsPayload.js
@@ -1,0 +1,56 @@
+export default [{
+  isComplete: false,
+  isLoading: false,
+  payload: {
+    webAppName: 'home',
+    webAppDescription: 'Home',
+    iconURL: '$home',
+    webAppLink: 'user-dashboard',
+    defaultSchema: '',
+    moduleName: 'apps',
+  },
+}, {
+  isComplete: false,
+  isLoading: false,
+  payload: {
+    webAppName: 'productionPlanning',
+    webAppDescription: 'Production Planning',
+    iconURL: '$productionPlanning',
+    webAppLink: 'production-planning',
+    defaultSchema: '[{"assetId":2,"primaryDisplayTag":"planid","partMatrixComposition":["part","machine","mold"]},{"assetId":3,"primaryDisplayTag":"planid","partMatrixComposition":["part","machine","tool"]}]',
+    moduleName: 'apps',
+  },
+}, {
+  isComplete: false,
+  isLoading: false,
+  payload: {
+    webAppName: 'productionLog',
+    webAppDescription: 'Production Log',
+    iconURL: '$productionLog',
+    webAppLink: 'production-log',
+    defaultSchema: '',
+    moduleName: 'apps',
+  },
+}, {
+  isComplete: false,
+  isLoading: false,
+  payload: {
+    webAppName: 'downtimeLog',
+    webAppDescription: 'Downtime Log',
+    iconURL: '$downtimeLog',
+    webAppLink: 'downtime-log',
+    defaultSchema: '',
+    moduleName: 'apps',
+  },
+}, {
+  isComplete: false,
+  isLoading: false,
+  payload: {
+    webAppName: 'liveShopfloor',
+    webAppDescription: 'Shopfloor Dashboard',
+    iconURL: '$shopfloorDashboard',
+    webAppLink: '/dashboards/#/d/live-shopfloor',
+    defaultSchema: '',
+    moduleName: 'dashboards',
+  },
+}];