+++ packages/services/api/element.service.js
@@ -13,8 +13,8 @@ class ElementService {
     return this.request.get(`/server/elements/site/${id}`);
   }
 
-  getRecordsByTags(payload) {
-    return this.request.post(`/server/elements/${payload.elementName}/getrecords${payload.queryParam}`, payload.request);
+  getRecordsByTags(elementName, queryParam, request) {
+    return this.request.post(`/server/elements/${elementName}/getrecords${queryParam}`, request);
   }
 
   postStreamRecords(payload) {
@@ -48,9 +48,17 @@ class ElementService {
   uploadFiles(elementName, payload) {
     return this.request.post(`/server/uploadfile/${payload.filecontent}/${payload.filename}?elementName=${elementName}&extension=${payload.fileextension}&assetId=${payload.assetid}`, '');
   }
+
+  downloadData(queryParam, payload) {
+    return this.request.post(`/server/data/download${queryParam}`, payload);
+  }
   
   getRecords(elementName, queryParam) {
     return this.request.get(`/server/elements/${elementName}/records${queryParam}`);
+  }
+  
+  getCollectionRecords(collectionName, queryParam) {
+    return this.request.get(`/server/collection/${collectionName}${queryParam}`);
   }
 
   getTrainingLogsRecords(jobId) {